+++ edit_form.php
@@ -24,9 +24,7 @@
  * @license       http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-if (!defined('MOODLE_INTERNAL')) {
-    die('Direct access to this script is forbidden.'); // It must be included from a Moodle page.
-}
+defined('MOODLE_INTERNAL') || die();
 
 require_once($CFG->libdir.'/formslib.php');
 
@@ -46,6 +44,8 @@ class edit_letter_form extends moodleform {
      * @throws coding_exception
      */
     public function definition() {
+        global $CFG;
+
         $mform            =&$this->_form;
         $id               = $this->_customdata['id'];
         $num              = $this->_customdata['num'];
@@ -57,12 +57,24 @@ class edit_letter_form extends moodleform {
         $actcoverage      = $this->_customdata['actcoverage'];
         $newcoverage      = $this->_customdata['newcoverage'];
 
+        $showgradeitemtypes = (isset($CFG->gradedist_showgradeitemtype)) ? $CFG->gradedist_showgradeitemtype : 0;
+
         $mform->addElement('header', 'gradedistheader', get_string('gradeletter', 'gradereport_gradedist'));
         $mform->addHelpButton('gradedistheader', 'pluginname', 'gradereport_gradedist');
 
         $select = $mform->createElement('select', 'gradeitem', get_string('gradeitem', 'gradereport_gradedist'));
         foreach ($gradeitems as $index => $gradeitem) {
-            $select->addOption($gradeitem->name, $index, ($gradeitem->disable) ? array( 'disabled' => 'disabled') : null);
+            $name = $gradeitem->name;
+            // If showgradeitemtype-setting is off property module is empty.
+            if ($showgradeitemtypes && $gradeitem->module) {
+                $modname = 'mod_'.$gradeitem->module;
+                $name .= " (".get_string('pluginname', $modname).")";
+            } else if ($showgradeitemtypes && $gradeitem->type == "manual") {
+                $name .= " (".get_string('manualitem', 'grades').")";
+            } else if ($gradeitem->type == get_string('gradecategory', 'grades')) {
+                $name .= " (".get_string('gradecategory', 'grades').")";
+            }
+            $select->addOption($name, $index, ($gradeitem->disable) ? [ 'disabled' => 'disabled' ] : null);
         }
         $mform->addElement($select);
 
@@ -80,44 +92,49 @@ class edit_letter_form extends moodleform {
             }
             $mform->addElement($selectgrouping);
         }
-        $gradeletters = array();
-        $gradeboundaries = array();
-        $gradeboundariesnew = array();
-        $attributes = array('style' => 'width:65px;margin-right:10px');
+        $gradeletters = [];
+        $gradeboundaries = [];
+        $gradeboundariesnew = [];
+        $attributes = ['style' => 'width:65px;margin-right:10px'];
 
         for ($i = 1; $i < $num + 1; $i++) {
             $gradeletters[] =& $mform->createElement('text', $i, false,
-                    array_merge(array('class' => 'gradeletters', 'disabled' => 'disabled'), $attributes));
+                    array_merge(['class' => 'gradeletters', 'disabled' => 'disabled'], $attributes));
             $gradeboundaries[] =& $mform->createElement('text', $i, false,
-                    array_merge(array('class' => 'gradeboundaries', 'disabled' => 'disabled'), $attributes));
+                    array_merge(['class' => 'gradeboundaries', 'disabled' => 'disabled'], $attributes));
             $gradeboundariesnew[] =& $mform->createElement('text', $i, false,
-                    array_merge(array('class' => 'gradeboundaries_new'), $attributes));
+                    array_merge(['class' => 'gradeboundaries_new'], $attributes));
         }
 
         $mform->addGroup($gradeletters, 'grp_gradeletters',
-                get_string('gradeletter', 'gradereport_gradedist'), array(''));
+                get_string('gradeletter', 'gradereport_gradedist'), ['']);
         $mform->setType('grp_gradeletters', PARAM_TEXT);
         $mform->addGroup($gradeboundaries, 'grp_gradeboundaries',
-                get_string('gradeboundary', 'gradereport_gradedist'), array(''));
+                get_string('gradeboundary', 'gradereport_gradedist'), ['']);
         $mform->setType('grp_gradeboundaries', PARAM_TEXT);
         $mform->addHelpButton('grp_gradeboundaries', 'gradeboundary', 'gradereport_gradedist');
         $mform->addGroup($gradeboundariesnew, 'grp_gradeboundaries_new',
-                get_string('gradeboundary_new', 'gradereport_gradedist'), array(''));
+                get_string('gradeboundary_new', 'gradereport_gradedist'), ['']);
         $mform->setType('grp_gradeboundaries_new', PARAM_TEXT);
         $mform->addHelpButton('grp_gradeboundaries_new', 'gradeboundary_new', 'gradereport_gradedist');
 
         $mform->addElement('html', '<div id="boundary_error_container"></div>');
 
+        // Buttons.
+        if ($edit) {
+            $mform->addElement('submit', 'submitbutton', get_string('changeletters', 'gradereport_gradedist'));
+        }
+
         $mform->addElement('header', 'chartheader', get_string('chart', 'gradereport_gradedist'));
 
-        $description = array();
+        $description = [];
         $description[] =& $mform->createElement('radio', 'description', '', get_string('absolut', 'gradereport_gradedist'), false);
         $description[] =& $mform->createElement('radio', 'description', '', get_string('percent', 'gradereport_gradedist'), true);
         $mform->setDefault('description', 0);
 
-        $mform->addGroup($description, 'grp_description', get_string('description', 'gradereport_gradedist'), array(''));
+        $mform->addGroup($description, 'grp_description', get_string('description', 'gradereport_gradedist'), ['']);
 
-        $columns = array();
+        $columns = [];
         $columns[] =& $mform->createElement('advcheckbox', 'actualcolumns', '',
                 get_string('actualcolumns', 'gradereport_gradedist'));
         $columns[] =& $mform->createElement('advcheckbox', 'newcolumns', '',
@@ -125,7 +142,7 @@ class edit_letter_form extends moodleform {
         $mform->setDefault('grp_columns[actualcolumns]', true);
         $mform->setDefault('grp_columns[newcolumns]', true);
 
-        $mform->addGroup($columns, 'grp_columns', get_string('columns', 'gradereport_gradedist'), array(''));
+        $mform->addGroup($columns, 'grp_columns', get_string('columns', 'gradereport_gradedist'), ['']);
 
         $mform->addElement('html', '<canvas height="100" id="chart_container" class="hey"></canvas>');
 
@@ -136,39 +153,36 @@ class edit_letter_form extends moodleform {
             html_writer::span(get_string('newcoverage', 'gradereport_gradedist')."&nbsp;", 'coveragetext')
                 .html_writer::span($newcoverage[0].'/'.$newcoverage[1].' ('.$newcoverage[2].'%)', 'newcoverage')));
 
+        $mform->addElement('html', '<div>&nbsp;</div>');
+
         // Hidden params.
         $mform->addElement('hidden', 'id', $id);
         $mform->setType('id', PARAM_INT);
         $mform->addElement('hidden', 'confirm', true);
         $mform->setType('confirm', PARAM_BOOL);
+        $mform->addElement('hidden', 'grp_export', '');
+        $mform->setType('grp_export', PARAM_TEXT);
+
+        $mform->addElement('header', 'downloadheader', get_string('download', 'gradereport_gradedist'));
 
         $mform->addElement('html', '<div class="grgd">'.
-            get_string('exportasimage', 'gradereport_gradedist').'&nbsp;&nbsp;&nbsp;'.
-            '<a href="#png" class="grgd_png">&nbsp;'.get_string('downloadpng', 'gradereport_gradedist').'&nbsp;</a>| '.
-            '<a href="#jpg" class="grgd_jpg">&nbsp;'.get_string('downloadjpeg', 'gradereport_gradedist').'&nbsp;</a>| '.
-            '<a href="#pdf" class="grgd_pdf">&nbsp;'.get_string('downloadpdf', 'gradereport_gradedist').'&nbsp;</a>| '.
-            '<a href="#print" class="grgd_print">&nbsp;'.get_string('printchart', 'gradereport_gradedist').'&nbsp;</a>'.
+            get_string('exportasimage', 'gradereport_gradedist').'&nbsp;'.
+            '&nbsp;<a href="#png" class="grgd_png">'.get_string('downloadpng', 'gradereport_gradedist').'</a>&nbsp;|'.
+            '&nbsp;<a href="#jpg" class="grgd_jpg">'.get_string('downloadjpeg', 'gradereport_gradedist').'</a>&nbsp;|'.
+            '&nbsp;<a href="#pdf" class="grgd_pdf">'.get_string('downloadpdf', 'gradereport_gradedist').'</a>&nbsp;|'.
+            '&nbsp;<a href="#print" class="grgd_print">'.get_string('printchart', 'gradereport_gradedist').'</a>'.
             '</div>'
         );
 
-        $mform->addElement('header', 'submitheader', get_string('submitanddownload', 'gradereport_gradedist'));
-
-        $export = array();
-        $exportformats = array(MTablePDF::OUTPUT_FORMAT_ODS     => 'ods',
-                               MTablePDF::OUTPUT_FORMAT_CSV_TAB => 'csv',
-                               MTablePDF::OUTPUT_FORMAT_XLSX     => 'xlsx');
-
-        $export[] =& $mform->createElement('select', 'exportformat', '', $exportformats);
-        $export[] =& $mform->createElement('submit', 'export', get_string('download', 'gradereport_gradedist'));
-        $mform->addGroup($export, 'grp_export', get_string('export', 'gradereport_gradedist'), array(''));
-        $mform->setDefault('grp_export[exportformat]', MTablePDF::OUTPUT_FORMAT_XLSX);
+        $mform->addElement('html', '<div class="grgd">'.
+            get_string('export', 'gradereport_gradedist').'&nbsp;'.
+            '&nbsp;<a href="#xlsx" class="grgd_xlsx">'.get_string('xlsx', 'gradereport_gradedist').'</a>&nbsp;|'.
+            '&nbsp;<a href="#ods" class="grgd_ods">'.get_string('ods', 'gradereport_gradedist').'</a>&nbsp;|'.
+            '&nbsp;<a href="#csv" class="grgd_csv">'.get_string('csv', 'gradereport_gradedist').'</a>'.
+            '</div>'
+        );
 
         $mform->setExpanded('chartheader');
-        $mform->setExpanded('submitheader');
 
-        // Buttons.
-        if ($edit) {
-            $mform->addElement('submit', 'submitbutton', get_string('changeletters', 'gradereport_gradedist'));
-        }
     }
 }